+++ packages/components/checkbox/src/Checkbox.stories.tsx
@@ -73,12 +73,16 @@ export const ControlledState: StoryFn = _args => {
   )
 }
 
-const intent = ['primary', 'success', 'alert', 'error'] as const
+const intent = ['primary', 'success', 'alert', 'error', 'info', 'neutral'] as const
 
 export const Intent: StoryFn = _args => (
   <>
     {intent.map(color => {
-      return <Checkbox intent={color}>{color}</Checkbox>
+      return (
+        <Checkbox key={color} intent={color}>
+          {color}
+        </Checkbox>
+      )
     })}
   </>
 )