+++ packages/components/icon/src/Icon.variants.tsx
@@ -17,7 +17,7 @@ export const iconVariants = cva(['fill-current'], {
       neutral: ['text-neutral'],
     }),
     size: makeVariants<'size', ['current', 'sm', 'md', 'lg']>({
-      current: ['current-font-size'],
+      current: ['u-current-font-size'],
       sm: ['w-sz-16', 'h-sz-16'],
       md: ['w-sz-24', 'h-sz-24'],
       lg: ['w-sz-32', 'h-sz-32'],