+++ packages/components/switch/src/Switch.stories.tsx
@@ -1,5 +1,6 @@
 import { Button } from '@spark-ui/button'
-import { EyeFill, EyeOffFill } from '@spark-ui/icons'
+import { EyeFill } from '@spark-ui/icons/dist/icons/EyeFill'
+import { EyeOffFill } from '@spark-ui/icons/dist/icons/EyeOffFill'
 import { Meta, StoryFn } from '@storybook/react'
 import { type ComponentProps, useState } from 'react'
 
@@ -33,7 +34,7 @@ const switchSizes: ComponentProps<typeof Switch>['size'][] = ['sm', 'md']
 export const Sizes: StoryFn = _args => (
   <div className="gap-lg flex flex-col">
     {switchSizes.map(size => (
-      <Switch name="small" size={size}>
+      <Switch key={size} name="small" size={size}>
         {size}
       </Switch>
     ))}
@@ -53,7 +54,7 @@ const switchColors: ComponentProps<typeof Switch>['intent'][] = [
 export const Colors: StoryFn = _args => (
   <div className="gap-lg flex flex-col">
     {switchColors.map(color => (
-      <Switch name="primary" intent={color} defaultChecked>
+      <Switch key={color} name="primary" intent={color} defaultChecked>
         {color}
       </Switch>
     ))}