+++ packages/components/icon/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.10](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.5.9...@spark-ui/icon@1.5.10) (2023-05-02)
+
+**Note:** Version bump only for package @spark-ui/icon
+
+## [1.5.9](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.5.8...@spark-ui/icon@1.5.9) (2023-04-28)
+
+**Note:** Version bump only for package @spark-ui/icon
+
+## [1.5.8](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.5.7...@spark-ui/icon@1.5.8) (2023-04-27)
+
+### Bug Fixes
+
+- fixed warnings in stories ([d258423](https://github.com/adevinta/spark/commit/d258423e2dcc8f220b3cc6164fcb3bf6319106a8))
+
+## [1.5.7](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.5.6...@spark-ui/icon@1.5.7) (2023-04-26)
+
+**Note:** Version bump only for package @spark-ui/icon
+
+## [1.5.6](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.5.5...@spark-ui/icon@1.5.6) (2023-04-26)
+
+**Note:** Version bump only for package @spark-ui/icon
+
 ## [1.5.5](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.5.4...@spark-ui/icon@1.5.5) (2023-04-25)
 
 **Note:** Version bump only for package @spark-ui/icon