+++ packages/components/switch/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.5](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.10.4...@spark-ui/switch@1.10.5) (2023-05-02)
+
+### Bug Fixes
+
+- **switch:** icons import ([cca1aa4](https://github.com/adevinta/spark/commit/cca1aa43a44e91d7fc6095ddb54436286025e66e))
+
+## [1.10.4](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.10.3...@spark-ui/switch@1.10.4) (2023-04-28)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [1.10.3](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.10.2...@spark-ui/switch@1.10.3) (2023-04-27)
+
+### Bug Fixes
+
+- fixed warnings in stories ([d258423](https://github.com/adevinta/spark/commit/d258423e2dcc8f220b3cc6164fcb3bf6319106a8))
+
+## [1.10.2](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.10.1...@spark-ui/switch@1.10.2) (2023-04-27)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [1.10.1](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.10.0...@spark-ui/switch@1.10.1) (2023-04-26)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+# [1.10.0](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.9.2...@spark-ui/switch@1.10.0) (2023-04-26)
+
+### Features
+
+- **tailwind-plugins:** add new "shadow-border-transition" utility class ([289ff71](https://github.com/adevinta/spark/commit/289ff71d39e2c064132ed0d8c55e758a1ab21695)), closes [#676](https://github.com/adevinta/spark/issues/676)
+
 ## [1.9.2](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.9.1...@spark-ui/switch@1.9.2) (2023-04-25)
 
 **Note:** Version bump only for package @spark-ui/switch