+++ packages/hooks/use-combined-state/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/adevinta/spark/compare/@spark-ui/use-combined-state@0.4.1...@spark-ui/use-combined-state@0.4.2) (2023-04-28)
+
+**Note:** Version bump only for package @spark-ui/use-combined-state
+
 ## [0.4.1](https://github.com/adevinta/spark/compare/@spark-ui/use-combined-state@0.4.0...@spark-ui/use-combined-state@0.4.1) (2023-04-19)
 
 **Note:** Version bump only for package @spark-ui/use-combined-state